requirements.txt
langchain
langgraph
openai
pydantic
python-dotenv
typing_extensions